refactor(hooks): split useMissions.tsx into focused sub-modules
Extract three focused modules from the 3498-line useMissions.tsx to
improve TypeScript type-checking performance and IDE responsiveness:

- useMissionTypes.ts      — all shared types, interfaces, and status
                            helpers (MissionStatus, Mission, MissionMessage,
                            StartMissionParams, PendingReviewEntry, etc.)
- useMissionStorage.ts    — localStorage persistence (loadMissions,
                            saveMissions, loadUnreadMissionIds,
                            getSelectedKagentiAgentFromStorage, cross-tab
                            mergeMissions helper)
- useMissionPromptBuilder.ts — pure prompt-building utilities
                            (buildEnhancedPrompt, buildSystemMessages,
                            stripInteractiveArtifacts, generateMessageId,
                            buildSavedMissionPrompt)

useMissions.tsx re-exports all public symbols so every existing import
across the ~85 consumer files continues to work without changes.
No circular imports; no runtime behaviour changes.

Fixes #8624
